refactor(config): decouple Cloth Config GUI; introduce ConfigControllerRegistry
Previously, the config system had to be packaged with :cloth-config.
Now that is resolved, we can isolate the GUI by moving everything else
back to :common.

- ClothConfigScreenProvider now handles only the GUI, consuming controllers via registry
- Replaced ModConfigProvider / ClothConfigProvider SPI with ConfigControllerRegistry
  - Explicitly registers controller strategies instead of relying on service loader
- Moved SingletonModConfigController and ModConfigDTO to common module
- Prepares for optional GUI and cleaner domain separation
