+++ cloth-config/src/main/java/net/xolt/freecam/clothconfig/ModConfigScreenFactory.java
@@ -8,10 +8,10 @@ import me.shedaniel.clothconfig2.impl.builders.KeyCodeBuilder;
 import me.shedaniel.clothconfig2.impl.builders.SubCategoryBuilder;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.network.chat.Component;
-import net.xolt.freecam.config.model.ConfigController;
+import net.xolt.freecam.config.controller.ConfigController;
 import net.xolt.freecam.config.model.FlightMode;
 import net.xolt.freecam.config.ModBindings;
-import net.xolt.freecam.clothconfig.model.ModConfigDTO;
+import net.xolt.freecam.config.model.ModConfigDTO;
 import net.xolt.freecam.config.model.Perspective;
 
 import java.util.stream.Stream;
