+++ cloth-config/src/main/resources/META-INF/services/net.xolt.freecam.config.ConfigScreenProvider
@@ -1,1 +1,1 @@
-net.xolt.freecam.clothconfig.ClothConfigProvider
+net.xolt.freecam.clothconfig.ClothConfigScreenProvider
