--- cloth-config/src/main/resources/META-INF/services/net.xolt.freecam.config.ModConfigProvider
@@ -1,1 +1,0 @@
-net.xolt.freecam.clothconfig.ClothConfigProvider
