+++ common/src/main/java/net/xolt/freecam/config/MCAwareModConfig.java
@@ -1,6 +1,8 @@
 package net.xolt.freecam.config;
 
 import net.minecraft.world.level.block.Block;
+import net.xolt.freecam.config.controller.ConfigControllerRegistry;
+import net.xolt.freecam.config.model.ModConfigDTO;
 
 /**
  * Extends {@link ModConfig} with Minecraft-aware features.
@@ -8,7 +10,7 @@ import net.minecraft.world.level.block.Block;
 public interface MCAwareModConfig extends ModConfig {
 
     static MCAwareModConfig get() {
-        return ModConfigProvider.instance().getConfig();
+        return ConfigControllerRegistry.get(ModConfigDTO.class).getConfig();
     }
 
     // FIXME: interface should not use MC classes
